requirements.txt
Flask
yfinance
flask-restful
Pyrebase4
python-dotenv
flask-cors
yahooquery